fix subscribers list (#246)
Fixes: https://github.com/rabbitmq-community/rstream/issues/241


This PR refactors the consumer/subscriber system to use numeric IDs instead of string names for internal tracking, addressing issues with subscriber list management. The changes enhance performance and eliminate bugs associated with string-based subscriber identification.

- Converts subscriber tracking from string-based names to integer-based IDs for better performance and reliability
- Updates MessageContext and EventContext to include stream information directly instead of requiring lookups
- Adds validation for max_subscribers_by_connection limits and introduces new exception types


Breaking changes
======

1. Subscriber returns not the `subscriber_id`, but instead the`reference`

The bug was here [Tag 0.3.1]:
https://github.com/rabbitmq-community/rstream/blob/654a9ef23118d96098fe00861b3d661da7886030/rstream/consumer.py#L197-L203 

Given two references with the same name the  `subscriber = self._subscribers[reference]`  is not consistent.

With this PR the `_subscribers` is `[int, _Subscriber]`
```python
self._subscribers: dict[int, _Subscriber] = {}
```

Where the int is the subscriber id that _must_ be unique for connection by protocol.

2. remove the `get_stream` function `message_context.consumer.get_stream(message_context.subscriber_name)` 

The `get_stream` is not needed anymore since the `stream` is now passed on the `message_context` and also `event_context` 

3. `subscribe_name` is now optional
```python
class MessageContext:
    consumer: Consumer
    stream: str
    subscriber_id: int
    subscriber_name: Optional[str]
    offset: int
    timestamp: int
```
---------

Signed-off-by: Gabriele Santomaggio <G.santomaggio@gmail.com>